donor schema update - having an issue with user schema
