+++ src/types/donation.ts
@@ -2,7 +2,8 @@ import zBase from './base';
 import { z } from 'zod';
 import zObjectId from './objectId';
 import { zItemResponse } from './items';
-import { zDonorResponse, zUserResponse } from './persons';
+import { zUserResponse } from './persons';
+import { zDonorResponse } from './donors';
 
 export const evaluationEnum = ['High', 'Low', 'New'] as const;
