+++ src/types/persons.ts
@@ -22,28 +22,3 @@ export interface UserEntity extends z.infer<typeof zUserEntity> {}
 export interface CreateUserRequest extends z.infer<typeof zCreateUserRequest> {}
 
 export interface UserResponse extends z.infer<typeof zUserResponse> {}
-
-export const zDonorBase = zPersonBase.extend({
-  address: z.string(),
-  city: z.string(),
-  state: z.string(),
-  zip: z.number(),
-});
-
-export const zDonorEntity = zDonorBase.extend({ ...zBase.shape });
-
-export const zCreateDonorRequest = zDonorBase;
-
-export const zDonorResponse = zDonorEntity;
-
-export const zUpdateDonorRequest = zCreateDonorRequest.partial();
-
-export interface DonorEntity extends z.infer<typeof zDonorEntity> {}
-
-export interface CreateDonorRequest
-  extends z.infer<typeof zCreateDonorRequest> {}
-
-export interface DonorResponse extends z.infer<typeof zDonorResponse> {}
-
-export interface UpdateDonorRequest
-  extends z.infer<typeof zUpdateDonorRequest> {}
